+++ lib/phoenix/router/route.ex
@@ -10,32 +10,36 @@ defmodule Phoenix.Router.Route do
   The `Phoenix.Router.Route` struct. It stores:
 
     * :verb - the HTTP verb as an upcased string
+    * :kind - the kind of route, one of `:match`, `:forward`
     * :path - the normalized path as string
     * :host - the request host or host prefix
-    * :controller - the controller module
-    * :action - the action as an atom
+    * :plug - the plug module
+    * :opts - the plug options
     * :helper - the name of the helper as a string (may be nil)
     * :private - the private route info
     * :assigns - the route info
     * :pipe_through - the pipeline names as a list of atoms
 
   """
-  defstruct [:verb, :path, :host, :controller, :action,
+
+  defstruct [:verb, :kind, :path, :host, :plug, :opts,
              :helper, :private, :pipe_through, :assigns]
 
   @type t :: %Route{}
 
   @doc """
-  Receives the verb, path, controller, action and helper
+  Receives the verb, path, plug, options and helper
   and returns a `Phoenix.Router.Route` struct.
   """
-  @spec build(String.t, String.t, String.t | nil, atom, atom, atom | nil, atom, %{}, %{}) :: t
-  def build(verb, path, host, controller, action, helper, pipe_through, private, assigns)
-      when is_binary(verb) and is_binary(path) and (is_binary(host) or is_nil(host)) and
-           is_atom(controller) and is_atom(action) and (is_binary(helper) or is_nil(helper)) and
-           is_list(pipe_through) and is_map(private and is_map(assigns)) do
-    %Route{verb: verb, path: path, host: host, private: private,
-           controller: controller, action: action, helper: helper,
+  @spec build(:match | :forward, String.t, String.t, String.t | nil, atom, atom, atom | nil, atom, %{}, %{}) :: t
+  def build(kind, verb, path, host, plug, opts, helper, pipe_through, private, assigns)
+      when is_atom(verb) and (is_binary(host) or is_nil(host)) and
+           is_atom(plug) and (is_binary(helper) or is_nil(helper)) and
+           is_list(pipe_through) and is_map(private and is_map(assigns))
+           and kind in [:match, :forward] do
+
+    %Route{kind: kind, verb: verb, path: path, host: host, private: private,
+           plug: plug, opts: opts, helper: helper,
            pipe_through: pipe_through, assigns: assigns}
   end
 
@@ -43,20 +47,27 @@ defmodule Phoenix.Router.Route do
   Builds the expressions used by the route.
   """
   def exprs(route) do
-    {path, binding} = build_path_and_binding(route.path)
+    {path, binding} = build_path_and_binding(route)
 
     %{path: path,
       host: build_host(route.host),
+      verb_match: verb_match(route.verb),
       binding: binding,
       dispatch: build_dispatch(route, binding)}
   end
 
-  defp build_path_and_binding(path) do
-    {params, segments} = Plug.Router.Utils.build_path_match(path)
+  defp verb_match(:*), do: Macro.var(:_verb, nil)
+  defp verb_match(verb), do: verb |> to_string() |> String.upcase()
 
-    binding = Enum.map(params, fn var ->
+  defp build_path_and_binding(%Route{path: path} = route) do
+    {params, segments} = case route.kind do
+      :forward -> Plug.Router.Utils.build_path_match(path <> "/*_forward_path_info")
+      :match   -> Plug.Router.Utils.build_path_match(path)
+    end
+
+    binding = for var <- params, var != :_forward_path_info do
       {Atom.to_string(var), Macro.var(var, nil)}
-    end)
+    end
 
     {segments, binding}
   end
@@ -96,16 +107,57 @@ defmodule Phoenix.Router.Route do
     end
   end
 
-  defp build_pipes(route) do
-    initial = quote do
+  defp build_pipes(%Route{kind: :forward} = route) do
+    {_params, fwd_segments} = Plug.Router.Utils.build_path_match(route.path)
+
+    quote do
       var!(conn)
       |> Plug.Conn.put_private(:phoenix_pipelines, unquote(route.pipe_through))
       |> Plug.Conn.put_private(:phoenix_route, fn conn ->
-           opts = unquote(route.controller).init(unquote(route.action))
-           unquote(route.controller).call(conn, opts)
-         end)
-    end
-
+        opts = unquote(route.plug).init(unquote(route.opts))
+        Phoenix.Router.Route.forward(conn, unquote(fwd_segments), unquote(route.plug), opts)
+      end)
+    end |> pipe_through(route)
+  end
+  defp build_pipes(route) do
+    quote do
+      var!(conn)
+      |> Plug.Conn.put_private(:phoenix_pipelines, unquote(route.pipe_through))
+      |> Plug.Conn.put_private(:phoenix_route, fn conn ->
+        opts = unquote(route.plug).init(unquote(route.opts))
+        unquote(route.plug).call(conn, opts)
+      end)
+    end |> pipe_through(route)
+  end
+  defp pipe_through(initial, route) do
     Enum.reduce(route.pipe_through, initial, &{&1, [], [&2, []]})
+  end
+
+  @doc """
+  Forwards requests to another Plug at a new path.
+  """
+  def forward(%Plug.Conn{path_info: path, script_name: script} = conn, fwd_segments, target, opts) do
+    new_path = path -- fwd_segments
+    {base, ^new_path} = Enum.split(path, length(path) - length(new_path))
+    conn = %{conn | path_info: new_path, script_name: script ++ base} |> target.call(opts)
+    %{conn | path_info: path, script_name: script}
+  end
+
+  @doc """
+  Validates and returns the list of forward path segments.
+
+  Raises RuntimeError plug is already forwarded or path contains
+  a dynamic segment.
+  """
+  def forward_path_segments(path, plug, phoenix_forwards) do
+    case Plug.Router.Utils.build_path_match(path) do
+      {[], path_segments} ->
+        if phoenix_forwards[plug] do
+          raise ArgumentError, "`#{inspect plug}` has already been forwarded to. A module can only be forwarded a single time."
+        end
+        path_segments
+      _ ->
+        raise ArgumentError, "Dynamic segment `\"#{path}\"` not allowed when forwarding. Use a static path instead."
+    end
   end
 end