+++ test/mix/tasks/phoenix.routes_test.exs
@@ -11,6 +11,6 @@ defmodule Mix.Tasks.Phoenix.RoutesTest do
   test "format routes for specific router" do
     Mix.Tasks.Phoenix.Routes.run(["Mix.RouterTest"])
     assert_received {:mix_shell, :info, [routes]}
-    assert routes =~ "page_path  GET  /  PageController.index/2"
+    assert routes =~ "page_path  GET  /  PageController :index"
   end
 end